+++ src/classes/api/index.ts
@@ -1503,6 +1503,20 @@ export function getTimeConfiguration(calendarId: string = 'active'): SimpleCalen
 }
 
 /**
+ * If the calendar is open or not.
+ *
+ * @returns boolean True if the calendar is open, false if it is closed.
+ *
+ * @example
+ * ```javascript
+ * SimpleCalendar.api.isOpen(); // True or false depending on if the calendar is open or closed.
+ * ```
+ */
+export function isOpen(): boolean {
+    return MainApplication.rendered;
+}
+
+/**
  * Get if the current user is considered the primary GM or not.
  *
  * @returns If the current user is the primary GM.